Refactor work sidebar into feature modules
