+++ src/app/components/sidebar/work/index.ts
@@ -1,0 +1,1 @@
+export { WorkSidebarSection } from './work-sidebar-section'
