+++ src/app/components/sidebar/work/project-work-actions-menu.tsx
@@ -1,0 +1,69 @@
+import { FolderOpen, Pencil, Trash2 } from 'lucide-react'
+import { forwardRef, useState } from 'react'
+import type { DesktopActionInvoker } from '../../../desktop/types'
+import type { Project } from '../../../types'
+
+export const ProjectWorkActionsMenu = forwardRef<
+  HTMLDivElement,
+  {
+    project: Project
+    right: number
+    width: number
+    onAction: DesktopActionInvoker
+    onClose: () => void
+    onRename: () => void
+  }
+>(function ProjectWorkActionsMenuPanel(
+  { project, right, width, onAction, onClose, onRename },
+  ref,
+) {
+  const [deleteConfirm, setDeleteConfirm] = useState(false)
+  const runProjectAction = (action: 'project.open-in-file-manager' | 'project.remove-project') => {
+    void onAction(action, { projectId: project.id, projectName: project.name })
+    onClose()
+  }
+
+  return (
+    <div
+      ref={ref}
+      className="sidebar-work-project-actions-menu"
+      style={{ right: `${right}px`, width: `${width}px` }}
+      role="menu"
+    >
+      <button
+        type="button"
+        className="sidebar-work-project-actions-menu-item"
+        onClick={onRename}
+        role="menuitem"
+      >
+        <Pencil size={12} />
+        <span>Rename</span>
+      </button>
+      <button
+        type="button"
+        className="sidebar-work-project-actions-menu-item"
+        onClick={() => runProjectAction('project.open-in-file-manager')}
+        role="menuitem"
+      >
+        <FolderOpen size={12} />
+        <span>Reveal in file manager</span>
+      </button>
+      <button
+        type="button"
+        className="sidebar-work-project-actions-menu-item"
+        data-danger="true"
+        onClick={() => {
+          if (!deleteConfirm) {
+            setDeleteConfirm(true)
+            return
+          }
+          runProjectAction('project.remove-project')
+        }}
+        role="menuitem"
+      >
+        <Trash2 size={12} />
+        <span>{deleteConfirm ? 'Click to confirm' : 'Delete project'}</span>
+      </button>
+    </div>
+  )
+})
