+++ src/app/components/sidebar/work/work-sidebar-model.ts
@@ -1,0 +1,272 @@
+import type { ProjectGitState } from '../../../desktop/types'
+import type { Project, Thread } from '../../../types'
+
+const OLD_THREAD_THRESHOLD_MS = 7 * 24 * 60 * 60 * 1000
+export const UNASSIGNED_BRANCH_GROUP_ID = '__unassigned__'
+export function sameStringList(a: readonly string[], b: readonly string[]) {
+  return a.length === b.length && a.every((value, index) => value === b[index])
+}
+
+export function getProjectScopeLabel({
+  selectedProject,
+  visibleProjects,
+}: {
+  selectedProject: Project | null
+  visibleProjects: readonly Project[]
+}) {
+  if (visibleProjects.length === 0) return 'No projects selected'
+  const primaryVisibleProject =
+    visibleProjects.find((project) => project.id === selectedProject?.id) ?? visibleProjects[0]
+  const primaryProjectName = primaryVisibleProject?.name ?? selectedProject?.name ?? 'Projects'
+  return visibleProjects.length > 1
+    ? `${primaryProjectName} +${visibleProjects.length - 1}`
+    : primaryProjectName
+}
+
+export function getDisplayableProjects(projects: readonly Project[]) {
+  return projects.filter(
+    (project) => project.id.trim().length > 0 && project.name.trim().length > 0,
+  )
+}
+
+export function orderProjectsForScopeSelector(
+  projects: readonly Project[],
+  visibleProjectIds: string[],
+) {
+  const visibleIndexById = new Map(visibleProjectIds.map((projectId, index) => [projectId, index]))
+  return [...projects].sort((left, right) => {
+    const leftIndex = visibleIndexById.get(left.id)
+    const rightIndex = visibleIndexById.get(right.id)
+    if (leftIndex !== undefined && rightIndex !== undefined) return leftIndex - rightIndex
+    if (leftIndex !== undefined) return -1
+    if (rightIndex !== undefined) return 1
+    return 0
+  })
+}
+
+export function projectBlockMatchesSearch(input: {
+  branchThreads: Thread[]
+  currentBranch: string | null
+  normalizedSearchQuery: string
+  projectName: string
+  unassignedThreads: Thread[]
+}) {
+  if (input.normalizedSearchQuery.length === 0) return true
+  return (
+    input.projectName.toLowerCase().includes(input.normalizedSearchQuery) ||
+    (input.currentBranch?.toLowerCase().includes(input.normalizedSearchQuery) ?? false) ||
+    input.branchThreads.length > 0 ||
+    input.unassignedThreads.length > 0
+  )
+}
+
+export function getVisibleProjectIds(
+  storedVisibleProjectIds: string[] | null,
+  initialVisibleProjectIds: string[] | null | undefined,
+  selectedProject: Project | null,
+) {
+  if (storedVisibleProjectIds) return storedVisibleProjectIds
+  if (initialVisibleProjectIds) return initialVisibleProjectIds
+  if (storedVisibleProjectIds === null && initialVisibleProjectIds === null) {
+    return selectedProject ? [selectedProject.id] : []
+  }
+  return []
+}
+
+export type ThreadBuckets = {
+  activeThreads: Thread[]
+  olderThreads: Thread[]
+}
+
+export type BranchThreadGroup = {
+  id: string
+  label: string
+  threads: Thread[]
+  current: boolean
+  unassigned: boolean
+}
+
+export function getDirtyWorktreeMessage(
+  projectGitState: ProjectGitState | null,
+  projectId: string,
+) {
+  if (projectGitState?.projectId !== projectId || !projectGitState.isGitRepo) return null
+  const dirtyFileCount =
+    projectGitState.stagedFileCount +
+    projectGitState.unstagedFileCount +
+    projectGitState.untrackedFileCount
+  return dirtyFileCount > 0 ? 'Commit first' : null
+}
+
+function getThreadSortValue(thread: Thread) {
+  return thread.lastModifiedMs ?? 0
+}
+
+export function sortThreads(threads: Thread[]) {
+  return [...threads].sort((a, b) => getThreadSortValue(b) - getThreadSortValue(a))
+}
+
+export function bucketThreads(project: Project, selectedThreadId: string | null): ThreadBuckets {
+  const sortedThreads = sortThreads(project.threads)
+  const cutoffMs = Date.now() - OLD_THREAD_THRESHOLD_MS
+  const activeThreads: Thread[] = []
+  const olderThreads: Thread[] = []
+
+  for (const thread of sortedThreads) {
+    const shouldKeepVisible =
+      thread.id === selectedThreadId ||
+      Boolean(thread.pinned) ||
+      Boolean(thread.running) ||
+      Boolean(thread.unread) ||
+      (thread.lastModifiedMs ?? Number.MAX_SAFE_INTEGER) >= cutoffMs
+
+    if (shouldKeepVisible) {
+      activeThreads.push(thread)
+    } else {
+      olderThreads.push(thread)
+    }
+  }
+
+  return { activeThreads, olderThreads }
+}
+
+export function buildBranchGroups(
+  threads: Thread[],
+  currentBranch: string | null,
+  repositoryBranches: readonly string[],
+): BranchThreadGroup[] {
+  const groupedThreads = new Map<string, Thread[]>()
+  const unassignedThreads: Thread[] = []
+
+  for (const thread of threads) {
+    const branchName = thread.branchName?.trim()
+    if (!branchName) {
+      unassignedThreads.push(thread)
+      continue
+    }
+
+    const branchThreads = groupedThreads.get(branchName) ?? []
+    branchThreads.push(thread)
+    groupedThreads.set(branchName, branchThreads)
+  }
+
+  const branchNames = new Set<string>()
+  for (const branch of repositoryBranches) {
+    const normalizedBranch = branch.trim()
+    if (normalizedBranch) branchNames.add(normalizedBranch)
+  }
+  for (const branchName of groupedThreads.keys()) branchNames.add(branchName)
+  if (currentBranch) branchNames.add(currentBranch)
+
+  const groups: BranchThreadGroup[] = []
+  if (currentBranch && branchNames.has(currentBranch)) {
+    groups.push({
+      id: currentBranch,
+      label: currentBranch,
+      threads: sortThreads(groupedThreads.get(currentBranch) ?? []),
+      current: true,
+      unassigned: false,
+    })
+    branchNames.delete(currentBranch)
+  }
+
+  const otherBranchGroups = [...branchNames]
+    .map((branchName) => ({
+      id: branchName,
+      label: branchName,
+      threads: sortThreads(groupedThreads.get(branchName) ?? []),
+      current: false,
+      unassigned: false,
+    }))
+    .sort((a, b) => a.label.localeCompare(b.label))
+
+  groups.push(...otherBranchGroups)
+
+  if (unassignedThreads.length > 0) {
+    groups.push({
+      id: UNASSIGNED_BRANCH_GROUP_ID,
+      label: 'Unassigned',
+      threads: sortThreads(unassignedThreads),
+      current: false,
+      unassigned: true,
+    })
+  }
+
+  if (groups.length === 0) {
+    groups.push({
+      id: UNASSIGNED_BRANCH_GROUP_ID,
+      label: 'Unassigned',
+      threads: [],
+      current: false,
+      unassigned: true,
+    })
+  }
+
+  return groups
+}
+
+export function filterBranchGroups(groups: BranchThreadGroup[], searchQuery: string) {
+  const normalizedSearchQuery = searchQuery.trim().toLowerCase()
+  if (!normalizedSearchQuery) return groups
+
+  return groups
+    .map((group) => ({
+      ...group,
+      threads: group.threads.filter((thread) =>
+        [thread.title, thread.summary ?? '', thread.branchName ?? group.label]
+          .join(' ')
+          .toLowerCase()
+          .includes(normalizedSearchQuery),
+      ),
+    }))
+    .filter(
+      (group) =>
+        group.label.toLowerCase().includes(normalizedSearchQuery) || group.threads.length > 0,
+    )
+}
+
+export function getProjectGitStateForSidebar(
+  projectId: string,
+  projectGitState: ProjectGitState | null,
+  gitStatesByProjectId: ReadonlyMap<string, ProjectGitState | null>,
+) {
+  if (projectGitState?.projectId === projectId) return projectGitState
+  return gitStatesByProjectId.get(projectId) ?? null
+}
+
+export function getCurrentBranchForProject(
+  project: Project,
+  projectGitState: ProjectGitState | null,
+  gitStatesByProjectId: ReadonlyMap<string, ProjectGitState | null>,
+) {
+  const gitState = getProjectGitStateForSidebar(project.id, projectGitState, gitStatesByProjectId)
+  return gitState?.isGitRepo ? gitState.branch : null
+}
+
+export function getRepositoryBranchesForProject(
+  project: Project,
+  projectGitState: ProjectGitState | null,
+  gitStatesByProjectId: ReadonlyMap<string, ProjectGitState | null>,
+) {
+  const gitState = getProjectGitStateForSidebar(project.id, projectGitState, gitStatesByProjectId)
+  return gitState?.isGitRepo ? gitState.branches : []
+}
+
+export function filterThreadsForCurrentBranch(
+  threads: readonly Thread[],
+  currentBranch: string | null,
+) {
+  if (!currentBranch) return []
+  return sortThreads(threads.filter((thread) => thread.branchName === currentBranch))
+}
+
+export function filterThreadsBySearch(threads: readonly Thread[], searchQuery: string) {
+  const normalizedSearchQuery = searchQuery.trim().toLowerCase()
+  if (!normalizedSearchQuery) return [...threads]
+  return threads.filter((thread) =>
+    [thread.title, thread.summary ?? '', thread.branchName ?? '']
+      .join(' ')
+      .toLowerCase()
+      .includes(normalizedSearchQuery),
+  )
+}
