+++ src/app/components/sidebar/work/work-thread-row.tsx
@@ -1,0 +1,71 @@
+import type { DesktopActionInvoker } from '../../../desktop/types'
+import type { Project, Thread, View } from '../../../types'
+import { ThreadRow } from '../project-tree/thread-row'
+
+export function WorkThreadRow({
+  activeView,
+  project,
+  selectedThreadId,
+  terminalRunningSessionPaths,
+  thread,
+  currentBranch,
+  onAction,
+  onThreadOpen,
+}: {
+  activeView: View
+  project: Project
+  selectedThreadId: string | null
+  terminalRunningSessionPaths: ReadonlySet<string>
+  thread: Thread
+  currentBranch: string | null
+  onAction: DesktopActionInvoker
+  onThreadOpen: (projectId: string, threadId: string, sessionPath: string) => void
+}) {
+  const isSelected =
+    selectedThreadId === thread.id &&
+    (activeView === 'thread' || activeView === 'project' || activeView === 'gitops')
+
+  return (
+    <ThreadRow
+      age={thread.age}
+      pinned={Boolean(thread.pinned)}
+      running={Boolean(thread.running)}
+      terminalRunning={Boolean(
+        thread.sessionPath && terminalRunningSessionPaths.has(thread.sessionPath),
+      )}
+      unread={Boolean(thread.unread)}
+      isSelected={isSelected}
+      title={thread.title}
+      assignBranchLabel={
+        currentBranch && thread.branchName === currentBranch
+          ? `Unassign from ${currentBranch}`
+          : currentBranch
+            ? `Assign to ${currentBranch}`
+            : 'Clear assigned branch'
+      }
+      onDelete={() =>
+        onAction('thread.delete', {
+          projectId: project.id,
+          threadId: thread.id,
+        })
+      }
+      onAssignToBranch={() =>
+        onAction('thread.assign-branch', {
+          projectId: project.id,
+          threadId: thread.id,
+          branchName: currentBranch && thread.branchName === currentBranch ? null : currentBranch,
+        })
+      }
+      onOpen={() => {
+        if (!thread.sessionPath) return
+        onThreadOpen(project.id, thread.id, thread.sessionPath)
+      }}
+      onPin={() =>
+        onAction('thread.pin', {
+          projectId: project.id,
+          threadId: thread.id,
+        })
+      }
+    />
+  )
+}
